+++ Race_Element.HUD.ACC/Overlays/Driving/TrackMap/TrackMapCreationJob.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.IO;
-using System.Drawing;
 using System.Threading;
 using System.Diagnostics;
 using System.Collections.Generic;
@@ -10,125 +9,6 @@ using RaceElement.Data.ACC.Core;
 using RaceElement.Util;
 
 namespace RaceElement.HUD.ACC.Overlays.Driving.TrackMap;
-
-public class TrackPoint
-{
-    private float _x, _y, _spline;
-    private float _deltaTime, _kmh;
-    private float _accX, _accY, _accZ;
-
-    public TrackPoint()
-    {
-        _x = 0;
-        _y = 0;
-        _spline = 0;
-
-        _kmh = 0;
-        _deltaTime = 0;
-
-        _accX = 0;
-        _accY = 0;
-        _accZ = 0;
-    }
-
-    public TrackPoint(TrackPoint other)
-    {
-        _x = other._x;
-        _y = other._y;
-        _spline = other._spline;
-
-        _kmh = other._kmh;
-        _deltaTime = other._deltaTime;
-
-        _accX = other._accX;
-        _accY = other._accY;
-        _accZ = other._accZ;
-    }
-
-    public float X
-    {
-        get => _x;
-        set => _x = value;
-    }
-
-    public float Y
-    {
-        get => _y;
-        set => _y = value;
-    }
-
-    public float Spline
-    {
-        get => _spline;
-        set => _spline = value;
-    }
-
-    public float DeltaTime
-    {
-        get => _deltaTime;
-        set => _deltaTime = value;
-    }
-
-    public float Kmh
-    {
-        get => _kmh;
-        set => _kmh = value;
-    }
-
-    public float AccX
-    {
-        get => _accX;
-        set => _accX = value;
-    }
-
-    public float AccY
-    {
-        get => _accY;
-        set => _accY = value;
-    }
-
-    public float AccZ
-    {
-        get => _accZ;
-        set => _accZ = value;
-    }
-
-    public PointF ToPointF()
-    {
-        return new PointF(_x, _y);
-    }
-}
-
-public class TrackPointSplineComparator : IComparer<TrackPoint>
-{
-    private float _epsilon;
-
-    public TrackPointSplineComparator()
-    {
-        _epsilon = 0.0005f;
-    }
-
-    public TrackPointSplineComparator(float epsilon)
-    {
-        _epsilon = epsilon;
-    }
-
-    public int Compare(TrackPoint x, TrackPoint y)
-    {
-        if (x == null || y == null)
-        {
-            return 0;
-        }
-
-        var diff = x.Spline - y.Spline;
-        if (Math.Abs(diff) < _epsilon)
-        {
-            return 0;
-        }
-
-        return diff > 0 ? 1 : -1;
-    }
-}
 
 public class TrackMapCreationJob : AbstractLoopJob
 {
